pyproject.toml
[tool.poetry]
name = "room-booking-system"
version = "0.1.0"
description = ""
authors = ["tliangso <earth78203@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
whitenoise = "^6.6.0"
gunicorn = "^21.2.0"
hypercorn = "^0.15.0"
pillow = "^10.1.0"
psycopg2-binary = "^2.9.5"
django-ninja = "^1.0.1"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
